dataproc/v1/python/google-cloud-dataproc/v5/requirements.txt
google-cloud-dataproc==5.21.0
